+++ mlflow/genai/optimize/optimizers/__init__.py
@@ -1,0 +1,4 @@
+from mlflow.genai.optimize.optimizers.base_optimizer import _BaseOptimizer
+from mlflow.genai.optimize.optimizers.dspy_mipro_optimizer import _DSPyMIPROv2Optimizer
+
+__all__ = ["_BaseOptimizer", "_DSPyMIPROv2Optimizer"]
