+++ mlflow/genai/optimize/optimizers/dspy_mipro_optimizer.py
@@ -1,0 +1,124 @@
+import logging
+import math
+from typing import TYPE_CHECKING, Any, Optional, Union
+
+from mlflow.entities.model_registry import Prompt
+from mlflow.genai.optimize.optimizers.dspy_optimizer import _DSPyOptimizer
+from mlflow.genai.optimize.types import OBJECTIVE_FN, LLMParams
+from mlflow.genai.scorers import Scorer
+
+if TYPE_CHECKING:
+    import dspy
+    import pandas as pd
+
+
+_logger = logging.getLogger(__name__)
+
+
+class _DSPyMIPROv2Optimizer(_DSPyOptimizer):
+    def optimize(
+        self,
+        prompt: Prompt,
+        target_llm_params: LLMParams,
+        train_data: "pd.DataFrame",
+        scorers: list[Scorer],
+        objective: Optional[OBJECTIVE_FN] = None,
+        eval_data: Optional["pd.DataFrame"] = None,
+    ) -> Union[str, list[dict[str, Any]]]:
+        import dspy
+
+        _logger.info(f"Started optimizing prompt {prompt.uri}...")
+
+        input_fields = self._get_input_fields(train_data)
+        output_fields = self._get_output_fields(train_data)
+        signature = dspy.make_signature(
+            {
+                **{key: (_type, dspy.InputField()) for key, _type in input_fields.items()},
+                **{key: (_type, dspy.OutputField()) for key, _type in output_fields.items()},
+            },
+            prompt.template,  # TODO: Extract only instructions from the existing prompt template
+        )
+
+        # Define main student program
+        program = dspy.Predict(signature)
+
+        train_data = self._convert_to_dspy_dataset(train_data)
+        eval_data = self._convert_to_dspy_dataset(eval_data) if eval_data is not None else None
+
+        teacher_settings = {}
+        if self.optimizer_config.optimizer_llm:
+            teacher_lm = dspy.LM(
+                model=self.optimizer_config.optimizer_llm.model_name,
+                temperature=self.optimizer_config.optimizer_llm.temperature,
+                api_base=self.optimizer_config.optimizer_llm.base_uri,
+            )
+            teacher_settings["lm"] = teacher_lm
+
+        num_candidates = self.optimizer_config.num_instruction_candidates
+        optimizer = dspy.MIPROv2(
+            metric=self._convert_to_dspy_metric(input_fields, output_fields, scorers, objective),
+            max_bootstrapped_demos=self.optimizer_config.max_few_show_examples,
+            num_candidates=num_candidates,
+            num_threads=self.optimizer_config.num_threads,
+            teacher_settings=teacher_settings,
+            auto=None,
+        )
+
+        adapter = dspy.JSONAdapter()
+        lm = dspy.LM(
+            model=target_llm_params.model_name,
+            temperature=target_llm_params.temperature,
+            api_base=target_llm_params.base_uri,
+        )
+        with dspy.context(lm=lm, adapter=adapter):
+            dspy_logger = logging.getLogger("dspy")
+            original_level = dspy_logger.level
+            dspy_logger.setLevel(logging.ERROR)
+            try:
+                optimized_program = optimizer.compile(
+                    program,
+                    trainset=train_data,
+                    valset=eval_data,
+                    num_trials=self._get_num_trials(num_candidates),
+                    minibatch_size=self._get_minibatch_size(train_data, eval_data),
+                    requires_permission_to_run=False,
+                )
+            finally:
+                # Restore original logging level
+                dspy_logger.setLevel(original_level)
+
+            return self._format_optimized_prompt(
+                adapter=adapter,
+                program=optimized_program,
+                input_fields=input_fields,
+            )
+
+    def _get_num_trials(self, num_candidates: int) -> int:
+        # MAX(2*log(num_candidates), 3/2*num_candidates)
+        return int(max(2 * math.log2(num_candidates), 1.5 * num_candidates))
+
+    def _get_minibatch_size(
+        self,
+        train_data: "pd.DataFrame",
+        eval_data: Optional["pd.DataFrame"] = None,
+    ) -> int:
+        # The default minibatch size is 35 in MIPROv2.
+        if eval_data is not None:
+            return min(35, len(eval_data) // 2)
+        return min(35, len(train_data) // 2)
+
+    def _format_optimized_prompt(
+        self, adapter: "dspy.Adapter", program: "dspy.Predict", input_fields: dict[str, type]
+    ) -> str:
+        messages = adapter.format(
+            signature=program.signature,
+            demos=program.demos,
+            inputs={key: "{{" + key + "}}" for key in input_fields.keys()},
+        )
+
+        return "\n\n".join(
+            [
+                f"<{message['role']}>\n{message['content']}\n</{message['role']}>"
+                for message in messages
+            ]
+        )
